Refactor command option imports
Rename `options.ts` to `config.ts` across command modules for better clarity. Update imports and references to reflect the new file name. Introduce a `command` constant in each config file to streamline command usage in `yargs`. This refactor improves code organization and readability without altering functionality.
