+++ src/commands/changelog/index.ts
@@ -1,9 +1,9 @@
 import commandExecutor from '../../lib/utils/commandExecutor'
+import { builder, command, options } from './config'
 import { handler } from './handler'
-import { builder, options } from './options'
 
 export default {
-  command: 'changelog',
+  command,
   desc: 'Generate a changelog from current or target branch or provided commit range.',
   builder,
   handler: commandExecutor(handler),
