+++ src/commands/recap/config.ts
@@ -1,5 +1,4 @@
 import { Arguments, Argv, Options } from 'yargs'
-import recap from '.'
 import { getCommandUsageHeader } from '../../lib/ui/helpers'
 import { BaseCommandOptions } from '../types'
 
@@ -15,6 +14,8 @@ export type RecapArgv = Arguments<RecapOptions>
 export interface RecapLlmResponse {
   summary: string
 }
+
+export const command = 'recap'
 
 /**
  * Command line options via yargs
@@ -47,5 +48,5 @@ export const options = {
 } as Record<string, Options>
 
 export const builder = (yargs: Argv) => {
-  return yargs.options(options).usage(getCommandUsageHeader(recap.command))
+  return yargs.options(options).usage(getCommandUsageHeader(command))
 }
