+++ src/commands/recap/index.ts
@@ -1,9 +1,9 @@
 import commandExecutor from '../../lib/utils/commandExecutor'
+import { builder, command, options } from './config'
 import { handler } from './handler'
-import { builder, options } from './options'
 
 export default {
-  command: 'recap',
+  command,
   desc: 'Summarize the changes in the repository over a specified timeframe.',
   builder,
   handler: commandExecutor(handler),
