+++ src/commands/review/index.ts
@@ -1,9 +1,9 @@
 import commandExecutor from '../../lib/utils/commandExecutor'
+import { builder, command, options } from './config'
 import { handler } from './handler'
-import { builder, options } from './options'
 
 export default {
-  command: 'review',
+  command,
   desc: 'Review the staged changes',
   builder,
   handler: commandExecutor(handler),
