+++ src/index.ts
@@ -6,11 +6,11 @@ import init from './commands/init'
 import recap from './commands/recap'
 import review from './commands/review'
 
-import { ChangelogOptions } from './commands/changelog/options'
-import { CommitOptions } from './commands/commit/options'
-import { InitOptions } from './commands/init/options'
-import { RecapOptions } from './commands/recap/options'
-import { ReviewOptions } from './commands/review/options'
+import { ChangelogOptions } from './commands/changelog/config'
+import { CommitOptions } from './commands/commit/config'
+import { InitOptions } from './commands/init/config'
+import { RecapOptions } from './commands/recap/config'
+import { ReviewOptions } from './commands/review/config'
 import { Config } from './lib/config/types'
 import * as types from './lib/types'
